+++ lib/$globj.d.ts
@@ -1,12 +1,11 @@
-/// <reference types="jquery" />
 /**
  * Global objects for general use
  */
 declare class $globj {
-    static readonly document: JQuery;
-    static readonly window: JQuery;
-    static readonly html: JQuery;
-    static readonly body: JQuery;
-    static readonly htmlBody: JQuery;
+    static readonly document: JQuery<HTMLElement>;
+    static readonly window: JQuery<HTMLElement>;
+    static readonly html: JQuery<HTMLElement>;
+    static readonly body: JQuery<HTMLElement>;
+    static readonly htmlBody: JQuery<HTMLElement>;
 }
 export default $globj;