+++ src/stopBodyScrolling.ts
@@ -3,6 +3,7 @@
  * For LABOR.digital
  */
 import $globj from "./$globj";
+import {forEach} from "./forEach";
 
 declare global {
 	interface Window {
@@ -20,21 +21,21 @@ declare global {
 	}
 }
 
-if(typeof window.HELPERS_JS_STOP_BODY_SCROLLING === 'undefined') window.HELPERS_JS_STOP_BODY_SCROLLING = false;
-if(typeof window.HELPERS_JS_STOP_BODY_SCROLLING_POSITION === 'undefined')
+if (typeof window.HELPERS_JS_STOP_BODY_SCROLLING === "undefined") window.HELPERS_JS_STOP_BODY_SCROLLING = false;
+if (typeof window.HELPERS_JS_STOP_BODY_SCROLLING_POSITION === "undefined")
 	window.HELPERS_JS_STOP_BODY_SCROLLING_POSITION = 0;
 
 /**
  * Helper to prevent the body from being scrolled with a fix for the ios 9 safari which is a pain...
  * @param {boolean} state True to stop the scrolling, false to reenable it.
  */
-export function stopBodyScrolling(state){
+export function stopBodyScrolling(state) {
 	let isIos = !!navigator.platform && /iPad|iPhone|iPod/.test(navigator.platform);
 
 	// Stop the scrolling
-	if(state === true){
+	if (state === true) {
 		// Ignore if already stopped
-		if(window.HELPERS_JS_STOP_BODY_SCROLLING === true) return;
+		if (window.HELPERS_JS_STOP_BODY_SCROLLING === true) return;
 
 		// Mark as stopped
 		window.HELPERS_JS_STOP_BODY_SCROLLING = true;
@@ -43,32 +44,48 @@ export function stopBodyScrolling(state){
 		let oldBodyWidth = $globj.body.width();
 
 		// Disable scrolling
-		if(isIos){
+		if (isIos) {
 			// Store old position
 			window.HELPERS_JS_STOP_BODY_SCROLLING_POSITION = $globj.htmlBody.scrollTop();
 			// Stop body scrolling with extra ios fix...
-			$globj.htmlBody.css({'overflow': 'hidden', 'position': 'relative', 'height': '100%', 'transition': 'none'});
+			$globj.htmlBody.css({"overflow": "hidden", "position": "relative", "height": "100%", "transition": "none"});
 		} else {
-			$globj.body.css({'overflow': 'hidden', 'transition': 'none'});
+			$globj.body.css({"overflow": "hidden", "transition": "none"});
 		}
 
 		// Calculate difference
 		let bodyWidthDiff = $globj.body.width() - oldBodyWidth;
-		$globj.body.css({'padding-right': bodyWidthDiff});
-		$('*[data-fixed-body-scrolling-target]').css({'margin-left': -bodyWidthDiff});
+		$globj.body.css({"padding-right": bodyWidthDiff});
+		forEach($("*[data-fixed-body-scrolling-target]"), $o => {
+			const type = $o.dataFallback("fixed-body-scrolling-target", "");
+			switch (type) {
+				case "":
+					$o.css({"margin-left": -bodyWidthDiff});
+					break;
+				case "half":
+					$o.css({"margin-left": (-bodyWidthDiff) / 2});
+					break;
+				case "right":
+					$o.css({"margin-right": bodyWidthDiff});
+					break;
+				case "rightHalf":
+					$o.css({"margin-right": (bodyWidthDiff) / 2});
+					break;
+			}
+		});
 	} else {
 		// Ignore if already active
-		if(window.HELPERS_JS_STOP_BODY_SCROLLING === false) return;
+		if (window.HELPERS_JS_STOP_BODY_SCROLLING === false) return;
 
 		// Reenable body scrolling
-		if(isIos){
+		if (isIos) {
 			$globj.htmlBody
-				.css({'overflow': '', 'padding-right': '', 'transition': '', 'position': '', 'height': ''})
+				.css({"overflow": "", "padding-right": "", "transition": "", "position": "", "height": ""})
 				.scrollTop(window.HELPERS_JS_STOP_BODY_SCROLLING_POSITION);
 		} else {
-			$globj.htmlBody.css({'overflow': '', 'padding-right': '', 'transition': ''});
+			$globj.htmlBody.css({"overflow": "", "padding-right": "", "transition": ""});
 		}
-		$('*[data-fixed-body-scrolling-target]').css({'margin-left': ''});
+		$("*[data-fixed-body-scrolling-target]").css({"margin-left": "", "margin-right": ""});
 		// Mark as stopped
 		window.HELPERS_JS_STOP_BODY_SCROLLING = false;
 	}