+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Added additional options for the data-fixed-body-scrolling-target targets to make sure they can be adjusted in multiple ways
+### Fixed
+- Fixed the types of the $globj children to JQuery<HTMLElement>
+
 ## [1.1.8] - 2018-11-08
 ### Added
 - Added new ucFirst(), sanitizeString() and checkOnlineState() helpers